added dropdown menu and reorganized files into folders
